Split out a data point search controller
